Create order details screen
This screen wwill be used for creating AND visualizing an order

- add basic views
- add navigation with safe-args
- create OrderDetailsViewModel
- populate OrderDetailsFragment + create a Fragment extension to handle
  the back button
- create AddItemDialog to add items to a order

+++ app/src/main/java/com/carlosolimpio/minhasvendas/data/OrderRepositoryImpl.kt
@@ -35,7 +35,7 @@ class OrderRepositoryImpl(private val orderDao: OrderDao) : OrderRepository {
         orderDao.insertOrder(order.toOrderEntity())
     }
 
-    override suspend fun getOrderFromId(id: Int): Resource<Order?> {
+    override suspend fun getOrderFromId(id: Long): Resource<Order?> {
         val orderEntity = orderDao.getOrderFromId(id)
 
         return if (orderEntity != null) {
