+++ app/src/main/java/com/carlosolimpio/minhasvendas/domain/order/OrderRepository.kt
@@ -6,6 +6,6 @@ interface OrderRepository {
     suspend fun retrieveOrderId(): Resource<Long>
     suspend fun deleteOrderId(id: Long): Resource<Boolean>
     suspend fun saveOrder(order: Order)
-    suspend fun getOrderFromId(id: Int): Resource<Order?>
+    suspend fun getOrderFromId(id: Long): Resource<Order?>
     suspend fun getAllOrders(): Resource<List<Order>>
 }
